Fix all_permissions function path
The partial all_permissions is pointing to ('modules/permission/lib/queries/permission/get_all') is missing. Changing the path to 'modules/permission/get_permissions' seems to fix it. Although I think it may be better to instead add queries, commands, and hook folders in the permission module, and keep the original path as is.