Validate input before user creation (#36)
* Validate input before user creation

All input required for user registration must be validated first, before
proceeding with the action. Because in case of errors in the way,
actions can't be rolled back (not in a ACID transaction).

* Handle profile creation error

Return the merged value from hooks results after executing the user
create action. This delegates the responsibility of error handling to
the caller (in this case app/views/pages/users/register.liquid).